Added Exceptions

+++ db/db_user.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm.session import Session
 from db.models import DBUser
 from db.hash import Hash
-
+from fastapi import HTTPException, status
 from models.user.schemas import UserBase
 
 
@@ -22,11 +22,20 @@ def get_all_users(db: Session):
 
 
 def get_user(db: Session, id: int):
-    return db.query(DBUser).filter(DBUser.id == id).first()
+    user = db.query(DBUser).filter(DBUser.id == id).first()
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found"
+        )
+    return user
 
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DBUser).filter(DBUser.id == id)
+    if not user.first():
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found"
+        )
     user.update(
         {
             DBUser.username: request.username,
@@ -40,6 +49,10 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DBUser).filter(DBUser.id == id).first()
+    if not user.first():
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found"
+        )
     db.delete(user)
     db.commit()
     return "ok"
